Improve float casting consistency among EBS scripts

+++ Assets/Scripts/Game/Bullet/Enemy bullet systems/Aries/AriesBulletSystem1.cs
@@ -14,27 +14,27 @@ public class AriesBulletSystem1 : EnemyBulletSystem
             {
                 for (int j = 0; j < 3; j++)
                 {
-                    float z = (i * 45) + (j * 120);
+                    int z = (i * 45) + (j * 120);
                     SpawnBullet(0, z, transform.up.RotateVectorBy(z));
                 }
 
                 yield return WaitForSeconds(ShootingCooldown);
             }
 
-            yield return WaitForSeconds(ShootingCooldown * 2);
+            yield return WaitForSeconds(ShootingCooldown * 2f);
 
             for (int i = 0; i < 3; i++)
             {
                 for (int j = 0; j < 30; j++)
                 {
-                    float z = j * 12 + (i * 6);
+                    int z = j * 12 + (i * 6);
                     SpawnBullet(1, z, Vector2.zero);
                 }
 
-                yield return WaitForSeconds(ShootingCooldown * 5);
+                yield return WaitForSeconds(ShootingCooldown * 5f);
             }
 
-            yield return ownerShip.MoveToRandomPosition(1, 1);
+            yield return ownerShip.MoveToRandomPosition(1f, 1f);
         }
     }
 }
